Bump log-cache submodule to contain upstream 2.1.1+fixes
Submodule's upstream 2.1.1 tag contains only 1 of the 2 commits we require, so
our fork now includes all of upstream's 2.1.1 plus the missing commit. Pull
request #2 on alphagov/paas-log-cache has details for this change.

When we deploy a version of CF > 7.5.0, which includes log-cache >= 2.2.0, we
can discard this release fork. Log-cache 2.2.0 contains the additional commit
we need.